Move testharness dispatch into testharness package
